% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.data.conv.type.parq.R
\name{def.data.conv.type.parq}
\alias{def.data.conv.type.parq}
\title{Convert data types in data frame to be compatible with parquet or avro schema data types}
\usage{
def.data.conv.type.parq(data, type, log = NULL)
}
\arguments{
\item{data}{Data frame of data to be converted.}

\item{type}{A data frame of: \cr
\code{name} Character. Name of data column
\code{type} Character. Data type. There may be multiple types, delimited by pipes (|)
\code{nullable} Boolean. Whether the field values are allowed to be null or NA}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
The input data frame with data types converted as specified
}
\description{
Definition function. Convert data types in data frame to match parquet or avro schema types.
}
\examples{
data <- data.frame(x=c(1,2,3),y=c('one','two','three'),stringsAsFactors=FALSE)
type <- data.frame(name=c('x'),type=c('string|utf8'),nullable=TRUE,stringsAsFactors=FALSE)
dataOut <- NEONprocIS.base::def.data.conv.type.parq(data=data,type=type)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\link[NEONprocIS.base]{def.write.parq}

\link[arrow]{data-type}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
