% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.data.mapp.schm.avro.R
\name{def.data.mapp.schm.avro}
\alias{def.data.mapp.schm.avro}
\title{Map data to an output avro schema}
\usage{
def.data.mapp.schm.avro(data, schm, ConvType = FALSE, log = NULL)
}
\arguments{
\item{data}{Data frame with named columns.}

\item{schm}{json string of the output schema}

\item{ConvType}{Logical TRUE to attempt to convert the data type to that indicated in the json
schema. FALSE to attempt no type conversion. Default is FALSE. Note: POSIX conversion to
numeric will result in seconds since epoc 1970-01-01 00:00:00.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame that (hopefully) matches the output schema
}
\description{
Definition function. Rearrange and/or create dummy columns in a data frame to match the column
order of an output avro schema. Optionally convert data type to match output schema.
}
\examples{
NEONprocIS.base::def.data.mapp.schma.avro(
}
\references{
Currently none
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
