% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.data.mapp.schm.parq.R
\name{def.data.mapp.schm.parq}
\alias{def.data.mapp.schm.parq}
\title{Map data to a parquet schema}
\usage{
def.data.mapp.schm.parq(data, schm, ConvType = FALSE, log = NULL)
}
\arguments{
\item{data}{Data frame with named columns.}

\item{schm}{A Parquet schema of class ArrowObject}

\item{ConvType}{Logical TRUE to attempt to convert the data type to that indicated in the 
schema. FALSE to attempt no type conversion. Default is FALSE.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame that (hopefully) matches the output schema
}
\description{
Definition function. Rearrange and/or create dummy columns in a data frame to match the column
order of a parquet schema. Optionally convert data type to match the schema.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
