% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.dir.in.R
\name{def.dir.in}
\alias{def.dir.in}
\title{Determine input directories based on expected subdirectories}
\usage{
def.dir.in(DirBgn, nameDirSub, log = NULL)
}
\arguments{
\item{DirBgn}{String. Directory path considered to be zero or more parents above the input directories.}

\item{nameDirSub}{String vector. The exact names of expected direct child directories (does not need to 
be a complete list).}
}
\value{
A character vector of input directories (datums)
}
\description{
Definition function. Determine the input directories (one or more) considered to be individual 
datums based on matching the names of child directories.
}
\examples{
Currently none
def.dir.in(DirBgn='/scratch/pfs/proc_group/prt/27134/2019/01/01',nameDirSub=c('data','calibration'))
}
\references{
Currently none
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
