% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.dir.splt.pach.time.R
\name{def.dir.splt.pach.time}
\alias{def.dir.splt.pach.time}
\title{Split Pachyderm directory into component folders and interpret date embedded in directory structure}
\usage{
def.dir.splt.pach.time(dir, log = NULL)
}
\arguments{
\item{dir}{String. Directory path (often found as an environment variable named as the input repository), structured as 
follows: #/pfs/BASE_REPO/#/yyyy/mm/dd/#, where # indicates any number (including zero) of 
parent and child directories of any name, so long as they are not 'pfs' or recognizable as the '/yyyy/mm/dd' structure which 
indicates the 4-digit year, 2-digit month, and 2-digit day. If 'pfs' is absent, the function will
fall back to environment variable \code{RELATIVE_PATH_INDEX} (Python pathlib-style, zero-based index of the *repo child*).
\code{RELATIVE_PATH_INDEX} should point to the first directory within the repo (e.g., product type); the repo is inferred as its parent.
Example: / is index 0, pfs is 1, BASE_REPO is 2, prt is 3; set \code{RELATIVE_PATH_INDEX=3} to infer repo BASE_REPO.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
A named list:\cr
\itemize{
   \item dirSplt = Character vector of the directory parents and children split into separate character strings (parsed by /)\cr
   \item repo = Character. The repository name (the child directory of /pfs)\cr
   \item idxRepo = Numeric. The index within \code{dirSplt} indicating the position of \code{repo}\cr
   \item dirRepo = Character. The directory structure within \code{repo}.\cr
   \item time = POSIXct. The /yyyy/mm/dd date (GMT) embedded within the directory structure (if present). NULL if cannot be interpreted.  
}
}
\description{
Definition function. Parse the directory structure of a Pachyderm repository located in a Docker container into  
its components and read the date optionally embedded within the file path.
}
\examples{
Currently none
def.dir.splt.pach.time('/scratch/pfs/proc_group/prt/27134/2019/01/01')
}
\references{
Currently none
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
