% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.file.avro.comb.ts.R
\name{def.file.avro.comb.ts}
\alias{def.file.avro.comb.ts}
\title{Merge the contents of multiple avro files that share a common time variable.}
\usage{
def.file.avro.comb.ts(file, nameVarTime, log = NULL)
}
\arguments{
\item{file}{Character vector of full or relative file paths. Must be avro format.}

\item{nameVarTime}{Character value. The name of the time variable common across all files.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame with the merged file contents.
}
\description{
Definition function. Merge the contents of multiple avro files that share a common time 
variable. Other than the time variable, the files must contain different columns. Any missing
timestamps among the files will be filled with NA values.
}
\examples{
# Not run
file <- c('/path/to/file1.avro','/path/to/file2.avro')
NEONprocIS.base::def.file.avro.comb.ts(file=file,nameVarTime='readout_time')
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.base]{def.read.avro.deve}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
