% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.file.name.out.R
\name{def.file.name.out}
\alias{def.file.name.out}
\title{Construct an output file name from an input file name}
\usage{
def.file.name.out(nameFileIn, prfx = NULL, sufx = NULL, ext = NULL, log = NULL)
}
\arguments{
\item{nameFileIn}{Character value. Input file name serving as a basis for the output file name.}

\item{prfx}{Character value. String to add to the front of nameFileIn. Defaults to NULL.}

\item{sufx}{Character value. String to append to the end of nameFileIn. If there are any 
periods (.) in the input file name, Sufx will be added before the final period. 
Defaults to NULL.}

\item{ext}{Character value. String indicating the extension of the output file name. If 
there are any periods (.) in the input file name, Ext will replace the characters after the 
final period. If there are no periods in the file name. A period followed by Ext will be added 
to the end of the file name. Defaults to NULL.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
A character vector of input directories (datums)
}
\description{
Definition function. Construct an output file name from an input file name, adding a prefix, 
appendix, extension, or all of the above.
}
\examples{
NEONprocIS.base::def.file.name.out(nameFileIn='myFileName.json',prfx='Prefix_',sufx='_Suffix',ext='txt')
}
\references{
Currently none
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
