% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.grp.meta.R
\name{def.grp.meta}
\alias{def.grp.meta}
\title{Get metadata/properties from groups json file}
\usage{
def.grp.meta(NameFile, log = NULL)
}
\arguments{
\item{NameFile}{Filename (including relative or absolute path). Must be json format.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
#' A data frame with group metadata.
}
\description{
Definition function. Read group json file and return a data frame of metadata/properties 
for the group.
}
\examples{
# Not run
# NameFile <- "/scratch/pfs/group_loader/pressure-air/rel-humidity_HARV000060.json"
# grpMeta <- NEONprocIS.base::def.grp.meta(NameFile=NameFile)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
None
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{currently}
\keyword{none}
