% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.loc.filt.R
\name{def.loc.filt}
\alias{def.loc.filt}
\title{Filter named location information by date-time range for NEON instrumented systems sensors}
\usage{
def.loc.filt(
  NameFileIn,
  NameFileOut = NULL,
  TimeBgn,
  TimeEnd = NULL,
  Prop = "all",
  log = NULL
)
}
\arguments{
\item{NameFileIn}{Filename (including relative or absolute path). Must be json format.}

\item{NameFileOut}{Filename (including relative or absolute path). Must be json format. Defaults to
NULL, in which case only the filtered json will be returned in list format}

\item{TimeBgn}{POSIX timestamp of the start time (inclusive)}

\item{TimeEnd}{POSIX timestamp of the end time (non-inclusive). Defaults to NULL, in which case the
location information will be filtered for the exact time of TimeBgn}

\item{Prop}{character vector of the properties in the location file to retain. Defaults to 'all', 
in which all properties of the original file are retained. Include here the names of properties embedded
at the top level of each location install entry (e.g. "HOR","VER","Data Rate") as well as within the 
"properties" list of each location install entry (e.g. "context","locations","install_date","remove_date",
"name","site"), as well as within the location_properties entry within "locations" 
(properties:locations:features:properties:location_features list). 
By default, adding "locations" will include geolocation coordinates and offsets to reference locations but 
will not include the "location_properties" list. To include "location_properties", add both "locations" 
and "location_properties" to the Prop argument. 
Note that "location_properties" is not allowed without "locations". If it is included alone, a warning 
will be issued and locations will not be included in the output.
By default, specifying "location_properties" will include all location properties, which often changes 
in the database as new properties are added. To filter for a particular set of location_properties (if 
they exist), use the syntax "location_properties:<property>", where <property> is the specific item in 
the location_properties list to include. Do not include spaces before or after the ":". 
Example: Prop="location_properties:UTM Zone"}
}
\value{
A list of filtered location information. If NameFileOut is specified, the filtered location
information will also be writted to file in the same json format of NameFileIn
}
\description{
Definition function. Read named location information (including geolocation history) from JSON file
for NEON instrumented systems sensor and filter it for a date-time range
}
\examples{
TimeBgn <- base::as.POSIXct('2018-01-01',tz='GMT')
TimeEnd <- base::as.POSIXct('2018-01-02',tz='GMT')
NameFileIn <- 'locations.json'
NameFileOut <- 'filtered_locations.json'
loc <- def.loc.filt(NameFileIn,NameFileOut,TimeBgn,TimeEnd)

Filtering for some basics, geolocation history (locations)m and some location properties
Prop <- c("HOR","VER","install_date","remove_date","name","site","Data Rate","locations","location_properties:UTM Zone","location_properties:TimeTube")
loc <- def.loc.filt(NameFileIn,NameFileOut,TimeBgn,TimeEnd,Prop)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
