% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.loc.geo.hist.R
\name{def.loc.geo.hist}
\alias{def.loc.geo.hist}
\title{Get geolocation history from a sensor-based location file}
\usage{
def.loc.geo.hist(NameFile, log = NULL)
}
\arguments{
\item{NameFile}{Filename (including relative or absolute path). Must be json format.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A list of configured locations found in the file. Nested within the list element for each 
configured location is a list of geolocation history, one list element per location change. 
Further nested within the list element for each geolocation is a variable list of properties 
of that geolocation. Each geolocation property may also be a list of elements.
}
\description{
Definition function. Read sensor locations json file and return the geolocation history of
all the configured locations within the file.
}
\examples{
# Not run
# NameFile <- '~/pfs/aquatroll200_23614_locations.json'
# locGeoHist <- NEONprocIS.base::def.loc.meta(NameFile=NameFile)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\link[NEONprocIS.base]{def.loc.meta}

\link[NEONprocIS.base]{def.loc.filt}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{currently}
\keyword{none}
