% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.loc.meta.R
\name{def.loc.meta}
\alias{def.loc.meta}
\title{Get metadata/properties from sensor locations json file}
\usage{
def.loc.meta(NameFile, NameLoc = NULL, TimeBgn = NULL,
  TimeEnd = NULL, log = NULL)
}
\arguments{
\item{NameFile}{Filename (including relative or absolute path). Must be json format.}

\item{NameLoc}{Character value of the named location to restrict output to. Defaults to NULL, 
in which case no filtering is done for named location}

\item{TimeBgn}{POSIXct timestamp of the start time of interest (inclusive). Defaults to NULL, 
in which case no filtering is done for installed time range}

\item{TimeEnd}{POSIXct timestamp of the end time of interest (non-inclusive). Defaults to NULL, in 
which case the location information will be filtered for the exact time of TimeBgn.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{

}
\description{
Definition function. Read sensor locations json file and return a data frame of metadata/properties 
filtered for a selected named location and/or install time range of interest
}
\examples{
# Not run
# NameFile <- '/scratch/pfs/prt_calibrated_location_group/prt/2019/01/01/767/location/prt_767_locations.json'
# NameLoc <- 'CFGLOC100140'
# TimeBgn <- base::as.POSIXct('2019-01-01',tz='GMT)
# TimeEnd <- base::as.POSIXct('2019-01-02',tz='GMT)
# locMeta <- NEONprocIS.base::def.loc.meta(NameFile=NameFile,NameLoc=NameLoc,TimeBgn=TimeBgn,TimeEnd=TimeEnd)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\link[NEONprocIS.base]{def.read.avro.deve}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{currently}
\keyword{none}
