% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.r.type.from.arrow.type.R
\name{def.r.type.from.arrow.type}
\alias{def.r.type.from.arrow.type}
\title{Get R data type for a given Arrow data type}
\usage{
def.r.type.from.arrow.type(typeArrow, log = NULL)
}
\arguments{
\item{typeArrow}{Arrow data type object}

\item{log}{Optional. A logger object as produced by NEONprocIS.base::def.log.init to produce structured log 
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be 
created for use within this function.}
}
\value{
An R data type object
}
\description{
Definition function. Given an arrow data type object, return the equivalent R data type object.
}
\examples{

# Integer type → integer()
def.r.type.from.arrow.type(arrow::int32())
#> integer()

# Floating point type → numeric()
def.r.type.from.arrow.type(arrow::float64())
#> numeric()
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\link[arrow]{data-type}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
