% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.rcd.fix.miss.na.R
\name{def.rcd.fix.miss.na}
\alias{def.rcd.fix.miss.na}
\title{Fill in missing timeseries records and NA-out corrupt records}
\usage{
def.rcd.fix.miss.na(data, timeBad, valuBad = NA, log = NULL)
}
\arguments{
\item{data}{Data frame with at least two columns, one of which must be:
\code{readout_time} POSIXct timestamps}

\item{timeBad}{A data frame with one column:
\code{readout_time} POSIXct timestamps signifying missing or corrupt records}

\item{valuBad}{the value to fill the records with. Defaults to NA.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame with the same columns as \code{data}. Any timestamps contained in \code{timeBad}
but not found in \code{data} are filled in with other values as NA. Any timestamps found in both
\code{timeBad} and \code{data} have all values except readout_time set to NA.
}
\description{
Definition function. Given a set of timestamps indicating missing or corrupt records, fill in
missing or corrupt timeseries records with a specific value.
}
\examples{
NEONprocIS.base::def.rcd.fix.miss.na(nameFile=c('/path/to/file1.avro','/another/path/to/file2.avro'))
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.base]{def.rcd.miss.na}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
