% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.rcd.miss.na.R
\name{def.rcd.miss.na}
\alias{def.rcd.miss.na}
\title{Identify any inconsistent or corrupt timeseries records across a set of files}
\usage{
def.rcd.miss.na(fileData, log = NULL)
}
\arguments{
\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}

\item{nameFile}{Character vector of any length, consisting of the full or relative paths to all
the data files to evaluate together for consistent timestamps and any NA-containing records. The
files must be avro files containing tabular data. These files will be read by 
NEONprocIS.base::def.read.avro.deve, which will return a data frame. One column of the data frame
must be readout_time.}
}
\value{
A list of \code{timeAll} and \code{timeBad}, each data frames with one column:
\code{readout_time} POSIXct timestamps signifying... for timeAll) the union of all encountered 
timestamps across the set of input files, and for timeBad) inconsistent or corrupt (NA-containing) 
records across the set of input files
}
\description{
Definition function. Given a set of file paths containing timeseries data, identify any timestamps
that are not consistently present across all files. Also identify any records that contain NA 
values.
}
\examples{
NEONprocIS.base::def.rcd.miss.na(nameFile=c('/path/to/file1.avro','/another/path/to/file2.avro'))
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.base]{def.read.avro.deve}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
