% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.read.avro.deve.R
\name{def.read.avro.deve}
\alias{def.read.avro.deve}
\title{Read AVRO file (development library)}
\usage{
def.read.avro.deve(NameFile, NameLib = "ravro.so", log = NULL)
}
\arguments{
\item{NameFile}{String. Name (including relative or absolute path) of AVRO file.}

\item{NameLib}{String. Name (including relative or absolute path) of AVRO library.}
}
\value{
A data frame of the data contained in the AVRO file. The schema is included in attribute 'schema'
}
\description{
Definition function. Read in AVRO file. Uses a super developmental version of the library. The 
requisite dependent libraries must be installed on the host system.
}
\examples{
myData <- NEONprocIS.base::def.read.avro.deve(NameFile='/scratch/test/myFile.avro',NameLib='/ravro.so')
attr(myData,'schema') # Returns the schema
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
