% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.read.evnt.json.R
\name{def.read.evnt.json}
\alias{def.read.evnt.json}
\title{Read event data in json format}
\usage{
def.read.evnt.json(NameFile, log = NULL)
}
\arguments{
\item{NameFile}{String. Name (including relative or absolute path) of event json file.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log 
output in addition to standard R error messaging. Defaults to NULL, in which no logger other than 
standard R error messaging will be used.}
}
\value{
A data frame of the data contained in the json file.
}
\description{
Definition function. Read event data in json format into data frame. Converts timestamps to POSIX.
NOTE: The json records must have uniform format.
}
\examples{
myEventData <- NEONprocIS.base::def.read.evnt.json(NameFile='/scratch/test/myFile.json')
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\code{\link[NEONprocIS.base]{def.log.init}}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
