% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.read.parq.ds.R
\name{def.read.parq.ds}
\alias{def.read.parq.ds}
\title{Read in a parquet dataset}
\usage{
def.read.parq.ds(fileIn, Var = NULL, VarTime = NULL, Df = FALSE, log = NULL)
}
\arguments{
\item{fileIn}{Character vector. Names (including relative or absolute path) of parquet files. 
These files must all have the same parquet schema. This can also be a directory, in which case
all files in the directory will be read into the dataset.}

\item{Var}{Optional. Character vector. Names of the fields/variables in the parquet dataset
to retain, ordered the same as Var. Only those present in the dataset will be output.}

\item{VarTime}{Optional. Character value of the time variable to order by. If NULL, no ordering
is performed.}

\item{Df}{Optional. Boolean. TRUE to return a data frame with the combined dataset. Defaults to FALSE.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
If Df is TRUE, a dataframe of the combined output. Any time variables are converted to POSIXct
and converted to GMT. If Df is FALSE, an object of class arrow_dplyr_query, which can be further 
manipulated using dplyr. Note that no time zone conversion is performed for the arrow_dplyr_query
object.
}
\description{
Definition function. Read in multiple parquet files with the same schema into an arrow dataset. 
Optionally limit to selected variables and order by time.
}
\examples{
# NOT RUN
fileIn <- c('path/to/file1.parquet','path/to/file2.parquet')
data <- NEONprocIS.base::def.read.parq.ds(fileIn,VarTime='readout_time',Df=TRUE)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
