% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.read.parq.R
\name{def.read.parq}
\alias{def.read.parq}
\title{Read Parquet file}
\usage{
def.read.parq(NameFile, log = NULL)
}
\arguments{
\item{NameFile}{String. Name (including relative or absolute path) of AVRO file.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame of the data contained in the Parquet file. The schema is included in
attribute 'schema'.
}
\description{
Definition function. Read in AVRO file. Uses a super developmental version of the library. The 
requisite dependent libraries must be installed on the host system.
}
\examples{
myData <- NEONprocIS.base::def.read.parq(NameFile='/scratch/test/myFile.avro')
attr(myData,'schema') # Returns the schema
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
