% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.schm.avro.pars.R
\name{def.schm.avro.pars}
\alias{def.schm.avro.pars}
\title{Parse AVRO schema to list}
\usage{
def.schm.avro.pars(FileSchm = NULL, Schm = NULL, log = NULL)
}
\arguments{
\item{FileSchm}{String. Optional. Full or relative path to schema file. One of FileSchm or Schm must be 
provided.}

\item{Schm}{String. Optional. Json formatted string of the AVRO file schema. One of FileSchm or Schm must 
be provided. If both Schm and FileSchm are provided, Schm will be ignored.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log 
output in addition to standard R error messaging. Defaults to NULL, in which no logger other than 
standard R error messaging will be used.}
}
\value{
A list of:\cr
\code{schmJson}: the avro schema in json format\cr
\code{schmList}: a list of avro schema properties and fields
\code{var}: a data frame of variables/fields in the schema, their data type(s), and documentation string
}
\description{
Definition function. Turn the JSON formatted AVRO schema (either from file or a json string)
into a formatted list.
}
\examples{
NEONprocIS.base::def.schm.avro.pars(FileSchm='/pfs/avro_schemas/prt_calibrated.avsc')
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\link[NEONprocIS.cal]{def.read.avro.deve}

\link[NEONprocIS.base]{def.log.init}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
