% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.schm.parq.from.schm.avro.R
\name{def.schm.parq.from.schm.avro}
\alias{def.schm.parq.from.schm.avro}
\title{Create Parquet schema from Avro schema}
\usage{
def.schm.parq.from.schm.avro(FileSchm = NULL, Schm = NULL, log = NULL)
}
\arguments{
\item{FileSchm}{String. Optional. Full or relative path to schema file. One of FileSchm or Schm must be 
provided.}

\item{Schm}{String. Optional. Json formatted string of the AVRO file schema. One of FileSchm or Schm must 
be provided. If both Schm and FileSchm are provided, Schm will be ignored.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log 
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be 
created for use within this function.}
}
\value{
An Apache Arrow Schema object containing a parquet schema representation of the avro schema.
}
\description{
Definition function. Create a arrow schema object from an avro schema file or json string.
}
\examples{
Currently none
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\link[arrow]{data-type}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
