% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.schm.parq.pars.R
\name{def.schm.parq.pars}
\alias{def.schm.parq.pars}
\title{Parse a parquet schema to retrieve data types of table columns}
\usage{
def.schm.parq.pars(schm, log = NULL)
}
\arguments{
\item{schm}{Schema object of parquet file retrieved from e.g. arrow::read_parquet(...)$schema}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame of: \cr
\code{name} Character. Name of data column
\code{type} Characer. Data type
\code{nullable} Boolean. Whether the field values are allowed to be null or NA
}
\description{
Definition function. Get the data types of each data column from a parquet schema.
}
\examples{
# Not run
# schm <- arrow::read_parquet(file='/path/to/file',as_data_frame=FALSE)$schema
NEONprocIS.base::def.schm.parq.pars(schm)
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.base]{def.read.parq}

\link[NEONprocIS.base]{def.wrte.parq}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
