% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.time.bin.diff.R
\name{def.time.bin.diff}
\alias{def.time.bin.diff}
\title{Create sequence of time bins as a difftime object}
\usage{
def.time.bin.diff(WndwBin, WndwTime, log = NULL)
}
\arguments{
\item{WndwBin}{WndwAgr A difftime object indicating the bin size (e.g. base::as.difftime(30,units="mins"))}

\item{WndwTime}{A difftime object indicating the total length of time to bin (e.g. base::as.difftime(1,units='days'))}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log 
output in addition to standard R error messaging. Defaults to NULL, in which no logger other than 
standard R error messaging will be used.}
}
\value{
A list of:
timeBgnDiff A difftime object providing the start times for each bin as a time offset from zero
timeEndDiff A difftime object providing the end times for each bin as a time offset from zero
}
\description{
Definition function. Create sequence of time bins as a difftime object offset from zero. This allows the 
time differences to be added to an arbitrary start time.
}
\examples{
WndwBin <- base::as.difftime(30,units="mins")
WndwTime <- base::as.difftime(1,units='days')
NEONprocIS.base::def.time.bin.diff(WndwBin=WndwBin,WndwTime=WndwTime)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\code{\link[NEONprocIS.base]{def.log.init}}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
