% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.time.miss.R
\name{def.time.miss}
\alias{def.time.miss}
\title{Determine missing time periods}
\usage{
def.time.miss(TimeBgn, TimeEnd, timeFull, log = NULL)
}
\arguments{
\item{TimeBgn}{A POSIXct timestamp of the start date of interest (inclusive)}

\item{TimeEnd}{A POSIXct timestamp of the end date of interest (exclusive)}

\item{timeFull}{A data frame of non-overlapping time ranges considered covered, with columns: \cr
\code{timeBgn} POSIXct timestamps of the start date of the time range (inclusive)\cr
\code{timeEnd} POSIXct timestamps of the end date of the time range (exclusive)\cr}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame of missing time ranges:\cr
\code{timeBgn} POSIXct. The start date-time (inclusive) of the missing time range \cr
\code{timeEnd} POSIXct. The end date-time (exclusive) of the missing time range \cr
}
\description{
Definition function. Given overall beginning and ending times and a data frame of time ranges,
determine the time periods that are missing.
}
\examples{
TimeBgn <- base::as.POSIXct('2019-01-01',tz='GMT')
TimeEnd <- base::as.POSIXct('2019-01-10',tz='GMT')
timeFull <- base::data.frame(timeBgn=as.POSIXct(c('2019-01-02','2019-01-05'),tz='GMT'),
                             timeEnd=as.POSIXct(c('2019-01-03','2019-01-07'),tz='GMT'))
# NEONprocIS.base::def.time.miss(TimeBgn=TimeBgn,TimeEnd=TimeEnd,timeFull=timeFull)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
None currently
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{currently}
\keyword{none}
