% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.validate.dataframe.R
\name{def.validate.dataframe}
\alias{def.validate.dataframe}
\title{Validate data frame not empty or invalid}
\usage{
def.validate.dataframe(dfIn)
}
\arguments{
\item{dfIn}{Input data frame to be validated}
}
\value{
Boolean. True if the input data frame is not empty AND does not have invalid values. FALSE otherwise. \cr
}
\description{
Validate a data frame to check it is empty or has invalid values.
Returns True if not empty AND has vaild values. FALSE otherwise.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}
}
\author{
Mija Choi \email{choim@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
