% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.validate.json.R
\name{def.validate.json}
\alias{def.validate.json}
\title{validate if the JSON content is strictly valid}
\usage{
def.validate.json(jsonIn)
}
\arguments{
\item{jsonIn}{Input json to be validated}
}
\value{
Boolean. TRUE if the input json is not empty AND valid. FALSE otherwise. \cr
}
\description{
Validate an input json to check it is valid.
Returns FALSE if the json is empty OR invaild . TRUE otherwise.
}
\examples{
NEONprocIS.base::def.validate.json (NameFileIn)
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}
}
\author{
Mija Choi \email{choim@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
