% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.validate.json.schema.R
\name{def.validate.json.schema}
\alias{def.validate.json.schema}
\title{validate if the JSON content is valid against the json schema}
\usage{
def.validate.json.schema(jsonIn, jsonSchemaIn, log = NULL)
}
\arguments{
\item{jsonIn}{Input json to be validated}

\item{jsonSchemaIn}{Input json schema to be validated against}
}
\value{
Boolean. TRUE if the input json is valid against the schema. FALSE if not valid. \cr
}
\description{
Validate an input json against the schema to check it is valid.
Run validate.json before validate.json.schema
Returns TRUE if the json is vaild . FALSE if invalid or error.
}
\examples{
NEONprocIS.base::def.validate.json.schema (jsonIn, jsonSchemaIn)
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}
}
\author{
Mija Choi \email{choim@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
