% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.validate.list.R
\name{def.validate.list}
\alias{def.validate.list}
\title{Validate list not empty or invalid}
\usage{
def.validate.list(listIn, log = NULL)
}
\arguments{
\item{listIn}{Input list to be validated}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
Boolean. True if the input it is a list and is not empty. FALSE otherwise. \cr
}
\description{
Validate a list to check it is not a list or it is not empty.
Returns True if it is a list and is not empty. FALSE otherwise.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}
}
\author{
Mija Choi \email{choim@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
