% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.validate.vector.notEmpty.R
\name{def.validate.vector.notEmpty}
\alias{def.validate.vector.notEmpty}
\title{Validate vector not empty}
\usage{
def.validate.vector.notEmpty(vectIn)
}
\arguments{
\item{vectIn}{Input vector to be validated}
}
\value{
Boolean. True if the input vector is not empty. FALSE otherwise. \cr
}
\description{
Validate a numeric vector to check it is empty.
Returns True if not empty. FALSE otherwise.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}
}
\author{
Mija Choi \email{choim@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
