% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.validate.vector.R
\name{def.validate.vector}
\alias{def.validate.vector}
\title{Validate vector}
\usage{
def.validate.vector(vectIn, TestEmpty = TRUE, TestNumc = TRUE, log = NULL)
}
\arguments{
\item{vectIn}{Input vector to be validated}

\item{TestEmpty}{Boolean. TRUE to test whether the vector is empty (function will return FALSE if empty).}

\item{TestNumc}{Boolean. TRUE to test whether the vector is numeric or logical (function will return FALSE if not numeric or logical (which converts to 0 or 1)).}
}
\value{
Boolean. True if the input vector is not empty. FALSE otherwise. \cr
}
\description{
Validate a vector to check that it is a vector, with optional tests for empty and numeric.
Returns True if passes all checks. FALSE otherwise.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}
}
\author{
Mija Choi \email{choim@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
