% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.var.mapp.in.out.R
\name{def.var.mapp.in.out}
\alias{def.var.mapp.in.out}
\title{Map input variable names to output variable names}
\usage{
def.var.mapp.in.out(
  nameVarIn = base::character(0),
  nameVarOut = base::character(0),
  nameVarDfltSame = base::character(0),
  log = NULL
)
}
\arguments{
\item{nameVarIn}{String vector of input variable names. Defaults to zero-length character vector.}

\item{nameVarOut}{String vector of output variable names matched to the same relative position 
of nameVarIn. Defaults to zero-length character vector.}

\item{nameVarDfltSame}{String vector of variable names that are to be mapped the same in & out
if they are not indicated in nameVarIn. Defaults to zero-length character vector.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame of \cr
\code{nameVarIn} Character. Input variable names
\code{nameVarOut} Character. Output variable names
}
\description{
Definition function. Create a data frame that maps input variable names to output variable names
given vectors of each. Option to add additional variable names that are the same in and out.
}
\examples{
NEONprocIS.base::def.var.mapp.in.out(nameVarIn=c('varIn1','varIn2'),
                                     nameVarOut=c('nameOut1','nameOut2'),
                                     nameVarDfltSame=c('nameSame1','nameSame2'))
}
\references{
Currently none
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
