% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.vect.pars.one.many.R
\name{def.vect.pars.one.many}
\alias{def.vect.pars.one.many}
\title{Parse vectors into key:multi-value sets}
\usage{
def.vect.pars.one.many(listVect, NameList = c("key", "value"),
  Type = rep(base::class(listVect[[1]]), 2), log = NULL)
}
\arguments{
\item{listVect}{A list of vectors of class numeric or character. Each vector must be length 2 or greater.}

\item{NameList}{Character vector of length 2 indicating first the name to call the 
key in each output list, and then the name to call the value in each output list
Names default to key and value if not input.}

\item{Type}{A 2-element vector indicating the respective R data classes of the key and 
value. Defaults to the class of listVect.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
A list of lists numbering the length of  with two elements, the key and value(s), except named according to NameList
}
\description{
Definition function. Parse the elements of one or more vectors into key:multiple value sets
}
\examples{
# Vector input represents key:multi-value set
listVect <- list(c('key1','1.5','3','4','5'),c('key2','3'))
NameList <- c('MyKeys','MyValues')
Type <- c('character','numeric')
NEONprocIS.base::def.vect.pars.one.many(listVect=listVect,NameList=NameList,Type=Type)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\code{\link[NEONprocIS.base]{def.arg.pars}}

\code{\link[NEONprocIS.base]{def.vect.pars.pair}}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
