% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.wrte.avro.deve.R
\name{def.wrte.avro.deve}
\alias{def.wrte.avro.deve}
\title{Write AVRO file (development library)}
\usage{
def.wrte.avro.deve(
  data,
  NameFile,
  NameSchm = NULL,
  NameSpceSchm = NULL,
  Schm = NULL,
  NameFileSchm = NULL,
  NameLib = "ravro.so"
)
}
\arguments{
\item{data}{Data frame. Data to write to file.}

\item{NameFile}{String. Name (including relative or absolute path) of AVRO file.}

\item{NameSchm}{String. Optional. Schema name.}

\item{NameSpceSchm}{String. Optional. Schema namepace.}

\item{Schm}{String. Optional. Json formatted string of the AVRO file schema. Example:\cr
"{\"type\" : \"record\",\"name\" : \"ST\",\"namespace\" : \"org.neonscience.schema.device\",\"fields\" : [ {\"name\" :\"readout_time\",\"type\" : {\"type\" : \"long\",\"logicalType\" : \"timestamp-millis\"},\"doc\" : \"Timestamp of readout expressed in milliseconds since epoch\"}, {\"name\" : \"soilPRTResistance\",\"type\" : [ \"null\", \"float\" ],\"doc\" : \"DPID: DP0.00041.001 TermID: 1728 Units: ohm Description: Soil Temperature, Level 0\",\"default\" : null} ],\"__fastavro_parsed\" : true}"\cr
Defaults to NULL, at which point the schema will be constructed using a different input or the data frame}

\item{NameFileSchm}{String. Optional. A filename (include relative or aboslute path) of an avro 
schema file (.avsc format). Defaults to NULL, at which point the schema will be constructed using a different 
input or from the data frame}

\item{NameLib}{String. Name (including relative or absolute path) of AVRO library. Defaults to ./ravro.so.}
}
\value{
Numeric. 0 = successful write.
}
\description{
Definition function. Write AVRO file from data frame. Uses a super developmental version of the library. The 
requisite dependent libraries must be installed on the host system.
}
\examples{
Currently none
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\link[NEONprocIS.base]{def.read.avro.deve}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
