% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.loc.meta.comb.R
\name{wrap.loc.meta.comb}
\alias{wrap.loc.meta.comb}
\title{Combine location metadata read from asset and/or named location files}
\usage{
wrap.loc.meta.comb(
  NameFile,
  NameLoc = NULL,
  TimeBgn = NULL,
  TimeEnd = NULL,
  log = NULL
)
}
\arguments{
\item{NameFile}{String array of filenames (including relative or absolute paths) of location files. Must be json format.}

\item{NameLoc}{Character value of the named location to restrict output to. Defaults to NULL, 
in which case no filtering is done for named location}

\item{TimeBgn}{POSIXct timestamp of the start time of interest (inclusive). Defaults to NULL, 
in which case no filtering is done for installed time range. Note that 
no time filtering is performed for location-based location files, since there is no sensor install 
information included, and only one location is included in the location-based location files.}

\item{TimeEnd}{POSIXct timestamp of the end time of interest (non-inclusive). Defaults to NULL, in 
which case the location information will be filtered for the exact time of TimeBgn. Note that 
no time filtering is performed for location-based location files, since there is no sensor install 
information included, and only one location is included in the location-based location files.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A list, length equal to the number of named locations found in the set of input
files. Each named location list element contains a nested list of location properties.
Note that location properties for each named location are populated with the first non-empty value
found among the files.
}
\description{
Definition function. Read in and merge location metadata from asset-focused location
file(s) and/or named-location-focused file(s). Any geolocation information found in the files 
is included.
}
\examples{
# Not run
# NameFile <- c('prt_767_locations.json','CFGLOC100140.json')
# NameLoc <- 'CFGLOC100140'
# TimeBgn <- base::as.POSIXct('2019-01-01',tz='GMT)
# TimeEnd <- base::as.POSIXct('2019-01-02',tz='GMT)
# locMeta <- NEONprocIS.base::wrap.loc.meta.comb(NameFile=NameFile,NameLoc=NameLoc,TimeBgn=TimeBgn,TimeEnd=TimeEnd)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\link[NEONprocIS.base]{def.loc.meta}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{currently}
\keyword{none}
