% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.dir.copy.symb.R, R/def.file.copy.symb.R
\name{def.dir.copy.symb}
\alias{def.dir.copy.symb}
\title{Copy a directory with a symbolic link}
\usage{
def.dir.copy.symb(DirSrc, DirDest, log = NULL)

def.dir.copy.symb(DirSrc, DirDest, log = NULL)
}
\arguments{
\item{DirSrc}{String vector. Source directories. All files in these source directories will be copied to the 
destination directories.}

\item{DirDest}{String value or vector. Destination director(ies). If not of length 1, must be same length 
as DirDest, each index corresponding to the same index of DirDest. NOTE: DirDest should be the parent of the 
distination directories. For example, to create a link from source /parent/child/ to /newparent/child, 
DirSrc='/parent/child/' and DirDest='/newparent/'}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
No output from this function other than performing the intended action.

No output from this function other than performing the intended action.
}
\description{
Definition function. Copy a source directory to a destination directory with a symbolic link

Definition function. Copy a file in a source directory to a destination directory with a symbolic link
}
\examples{
def.dir.copy.symb(DirSrc='/scratch/pfs/proc_group/prt/27134/2019/01/01',DirDest='pfs/out/prt/27134/2019/01/01')
def.dir.copy.symb(DirSrc='/scratch/pfs/proc_group/prt/27134/2019/01/01',DirDest='pfs/out/prt/27134/2019/01/01')
}
\references{
Currently none

Currently none
}
\seealso{
Currently none

Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}

Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
