% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.dir.crea.R
\name{def.dir.crea}
\alias{def.dir.crea}
\title{Create output directories}
\usage{
def.dir.crea(DirBgn, DirSub, log = log)
}
\arguments{
\item{DirBgn}{String value. Starting path}

\item{DirSub}{String value or vector. Path to subdirectories to create relative to starting
directory}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
No output from this function other than performing the intended action.
}
\description{
Definition function. Create one or more output directories given a starting path and 1 or more relative paths
}
\examples{
Currently none
NEONprocIS.base::def.dir.crea(DirBgn='/scratch/pfs/proc_group/prt',DirSub=c('relative/path/to/new/dir','relative/path2/to/new/dir2')
}
\references{
Currently none
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
