% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.loc.filt.R
\name{def.loc.filt}
\alias{def.loc.filt}
\title{Filter named location information by date-time range for NEON instrumented systems sensors}
\usage{
def.loc.filt(NameFileIn, NameFileOut = NULL, TimeBgn, TimeEnd = NULL,
  log = NULL)
}
\arguments{
\item{NameFileIn}{Filename (including relative or absolute path). Must be json format.}

\item{NameFileOut}{Filename (including relative or absolute path). Must be json format. Defaults to
NULL, in which case only the filtered json will be returned in list format}

\item{TimeBgn}{POSIX timestamp of the start time (inclusive)}

\item{TimeEnd}{POSIX timestamp of the end time (non-inclusive). Defaults to NULL, in which case the
location information will be filtered for the exact time of TimeBgn}
}
\value{
A list of filtered location information. If NameFileOut is specified, the filtered location
information will also be writted to file in the same json format of NameFileIn
}
\description{
Definition function. Read named location information (including geolocation history) from JSON file
for NEON instrumented systems sensor and filter it for a date-time range
}
\examples{
TimeBgn <- base::as.POSIXct('2018-01-01',tz='GMT)
TimeEnd <- base::as.POSIXct('2018-01-02',tz='GMT)
NameFileIn <- 'locations.json'
NameFileOut <- 'filtered_locations.json'
loc <- def.loc.filt(NameFileIn,NameFileOut,TimeBgn,TimeEnd)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
