% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.schm.parq.from.df.R
\name{def.schm.parq.from.df}
\alias{def.schm.parq.from.df}
\title{Create Parquet schema auto-detected from data frame}
\usage{
def.schm.parq.from.df(df, log = NULL)
}
\arguments{
\item{df}{Data frame}

\item{log}{Optional. A logger object as produced by NEONprocIS.base::def.log.init to produce structured log 
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be 
created for use within this function.}
}
\value{
An Apache Arrow Schema object for the input data frame
}
\description{
Definition function. Create an arrow schema object auto-detected from data frame
}
\examples{
Currently none
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\link[arrow]{data-type}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
