% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.validate.dataframe.R
\name{def.validate.dataframe}
\alias{def.validate.dataframe}
\title{Validate data frame not empty or invalid}
\usage{
def.validate.dataframe(dfIn, TestNa = FALSE, TestNumc = FALSE,
  log = NULL)
}
\arguments{
\item{dfIn}{Input data frame to be validated}

\item{TestNa}{Boolean where TRUE results in testing for any NA values (resulting in FALSE output). Defaults to FALSE.}

\item{TestNumc}{Boolean where TRUE results in testing for any non-numeric values (resulting in FALSE output). Defaults to FALSE.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
Boolean. True if the input data frame is not empty AND does not have invalid values. FALSE otherwise. \cr
}
\description{
Validate a data frame to check it is empty or has invalid values.
Returns True if not empty AND has vaild values. FALSE otherwise.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}
}
\author{
Mija Choi \email{choim@battelleecology.org}
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
