% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.wrte.parq.cast.cols.R
\name{def.wrte.parq.cast.cols}
\alias{def.wrte.parq.cast.cols}
\title{Cast Arrow Table Columns to Match Schema}
\usage{
def.wrte.parq.cast.cols(data, schm, log = NULL)
}
\arguments{
\item{data}{An Arrow table object}

\item{schm}{Target Arrow schema}

\item{log}{Optional logger object. Defaults to NULL.}
}
\value{
The Arrow table with columns cast to match the schema
}
\description{
Internal function. Casts arrow table columns to match the types specified in a target schema.
Handles special cases like string to large_utf8 conversion.
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{internal}
